Add files via upload
fixed createWhiteNoise() function, added support for wave files to be used as buffers. Commented on a lot of things that either need to be fixed or rewritten or reworked somehow to use proper Python3 resources instead of the old JavaScript dependencies. Deciphered some more of the JS syntax and command meanings and made note of those.